Stop API docs from yanking the next page after you leave.
unlockScroll removed listeners but left the 400ms scroll-pin rAF running, so navigating away mid-expand snapped the new page back to the docs position.
